bug: 읽음처리 안됨
