+++ src/apis/chat/chatList.api.js
@@ -9,11 +9,3 @@ export const getChatList = async (page) => {
   });
   return response.data.data;
 };
-
-//채팅 읽음 처리
-export const postReadMessage = async ({ chatroomId }) => {
-  const response = await authApi.post(
-    generateApiPath(API_DOMAINS.READ_MESSAGE, { chatroomId }),
-  );
-  return response.data.data;
-};
