+++ src/apis/chat/messages.api.js
@@ -10,3 +10,11 @@ export const getChatMessages = async ({ chatroomId, page }) => {
   );
   return response.data.data;
 };
+
+//채팅 읽음 처리
+export const postReadMessage = async ({ chatroomId }) => {
+  const response = await authApi.post(
+    generateApiPath(API_DOMAINS.READ_MESSAGE, { chatroomId }),
+  );
+  return response.data.data;
+};
